fix(reward): 로또 상금의 총합이 `int` 범위를 넘어갈 수 있으므로 `long`으로 변경

+++ src/test/java/lotto/model/LottoRewardStatisticsTest.java
@@ -28,7 +28,7 @@ class LottoRewardStatisticsTest {
 
         LottoRewardStatistics statistics = new LottoRewardStatistics(counter);
         int spent = counter.total() * Lotto.PRICE;
-        int reward = statistics.winnerRewards();
+        long reward = statistics.winnerRewards();
         float expected = reward / (float) spent;
